+++ pkg/thing/interactors/update_data_test.go
@@ -107,7 +107,7 @@ var updateDataUseCases = []UpdateDataTestCase{
 		"data sensorId doesn't match with thing's schema",
 		"authorization-token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 1, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 1, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
@@ -122,7 +122,7 @@ var updateDataUseCases = []UpdateDataTestCase{
 		"error publishing message in client exchange",
 		"authorization-token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 0, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 0, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
@@ -137,7 +137,7 @@ var updateDataUseCases = []UpdateDataTestCase{
 		"message successfuly send to client exchange",
 		"authorization token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 0, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 0, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
